+++ lib/rorvswild/plugin/action_view.rb
@@ -1,11 +1,14 @@
+# frozen_string_literal: true
+
 module RorVsWild
   module Plugin
     class ActionView
       def self.setup
         return if @installed
         return unless defined?(::ActiveSupport::Notifications.subscribe)
-        ActiveSupport::Notifications.subscribe("render_partial.action_view", new)
-        ActiveSupport::Notifications.subscribe("render_template.action_view", new)
+        ActiveSupport::Notifications.subscribe("render_partial.action_view", plugin = new)
+        ActiveSupport::Notifications.subscribe("render_template.action_view", plugin)
+        ActiveSupport::Notifications.subscribe("render_collection.action_view", plugin)
         @installed = true
       end
 
@@ -15,10 +18,11 @@ module RorVsWild
 
       def finish(name, id, payload)
         RorVsWild::Section.stop do |section|
-          section.kind = "view".freeze
+          section.kind = "view"
           section.commands << RorVsWild.agent.locator.relative_path(payload[:identifier])
           section.file = section.command
           section.line = 1
+          section.calls = payload[:count] if payload[:count] # render collection
         end
       end
     end